drc;
Updated all TC_super.m and sub-files to define the directory names in TC_super.m and pass them into the respective subfiles.
Fixed bug in TC_stats_plot.m so that the input to the vector t_day(...) is rounded to the nearest integer; apparently when dt is a fraction, the argument input as the index to the vector is read as a floating point and thus it gives an error.