drc;
Updated TC_stats.m to fix: the while loop for determining the timescale to equilibrium didn't have an escape for the case where there is no 30-day period that satisfies the equilibration criteria (within 10% of equilibrium value, slope less than critical value), so it would crash once the loop tried to go past day 150.  I've added in an if statement to save the equilibration timescale as NaN in such a case.